Updated credits logic

+++ backend/internal/dto/user.go
@@ -1,8 +1,8 @@
 package dto
 
 type UserDTO struct {
-	ID      string  `json:"id"`
-	Name    string  `json:"name"`
-	Email   string  `json:"email"`
-	Credits float64 `json:"credits"`
+	ID      string `json:"id"`
+	Name    string `json:"name"`
+	Email   string `json:"email"`
+	Credits int    `json:"credits"`
 }
